# sql: rationalize the handling of comments and improve performance