--- a/pkg/sql/comprules/BUILD.bazel
+++ b/pkg/sql/comprules/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
         "//pkg/sql/compengine",
         "//pkg/sql/lexbase",
         "//pkg/sql/scanner",
+        "//pkg/sql/sem/catconstants",
     ],
 )