… pkg/bench/rttanalysis/testdata/benchmark_expectations
… pkg/cli/zip_test.go
… pkg/sql/comprules/testdata/completion_patterns/schemas
… pkg/sql/crdb_internal.go
… pkg/sql/information_schema.go
… pkg/sql/logictest/testdata/logic_test/crdb_internal
… pkg/sql/logictest/testdata/logic_test/crdb_internal_catalog
… pkg/sql/logictest/testdata/logic_test/information_schema
… pkg/sql/logictest/testdata/logic_test/pg_builtins
… pkg/sql/logictest/testdata/logic_test/pg_catalog
… pkg/sql/logictest/testdata/logic_test/table
… pkg/sql/opt/exec/execbuilder/testdata/explain
… pkg/sql/opt/testutils/testcat/create_table.go
… pkg/sql/opt/testutils/testcat/test_catalog.go
… pkg/sql/opt/testutils/testcat/vtable.go
… pkg/sql/opt/xform/testdata/external/activerecord
… pkg/sql/opt/xform/testdata/external/liquibase
… pkg/sql/opt/xform/testdata/external/navicat
… pkg/sql/opt/xform/testdata/external/pgjdbc
… pkg/sql/pg_catalog.go
… pkg/sql/pg_metadata_test.go
… pkg/sql/sem/catconstants/constants.go
… pkg/sql/vtable/BUILD.bazel
… pkg/sql/vtable/crdb_internal.go
… pkg/sql/vtable/pg_catalog.go